Progress move test to RepositoryIT
